cloud_function/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
requests